requirements.txt
# Core dependencies for web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data processing
pandas==2.2.0
openpyxl==3.1.2  # For Excel export support
tabulate==0.9.0  # For markdown table formatting

# CLI and UI
rich==13.7.0
click==8.1.7
questionary==2.0.1

# Configuration and validation
python-dotenv==1.0.1
validators==0.22.0
pyyaml==6.0.1  # For YAML configuration files

# Utilities
ratelimit==2.2.1
fake-useragent==1.4.0
chardet==5.2.0  # For encoding detection
urllib3==2.1.0  # URL parsing utilities